Fix some saws adn other fixes
