+++ include/entity.h
@@ -6,7 +6,7 @@
 #include "physics.h"
 #include "transform.h"
 #include "game_stat.h"
-#include "obstacle.h"
+#include "obstacles.h"
 
 // Tile Type
 #define EMPTY 0
