+++ src/stage1_state.c
@@ -9,7 +9,7 @@
 #include "level.h"
 #include "SFX.h"
 #include "ui.h"
-#include "obstacle.h"
+#include "obstacles.h"
 #include "physics.h"
 
 // Definisi state
@@ -29,7 +29,7 @@ void stage1_init()
   player = create_entity(100, 400, 32, 32, (SDL_Color){0, 0, 0, 255});
   init_game_stat(&game_stat);
   start_timer(&game_stat);
-  
+
   SDL_Renderer *renderer = get_game_instance()->renderer;
   show_stage_transition(renderer, 1);
 
@@ -37,10 +37,11 @@ void stage1_init()
   {
     play_music(stage1_bgm, INT32_MAX);
   }
-
   if (current_level == 9) {
     setup_level_saws(current_level);
-}
+  }
+  change_level(current_level);
+  initiate_player(player, 570, 330);
 }
 
 void stage1_handle_input(SDL_Event *event)
@@ -77,8 +78,15 @@ void stage1_update(double delta_time)
 
   if (is_exit(&player->transform))
   {
-    
     SDL_Renderer *renderer = get_game_instance()->renderer;
+
+    if (current_level == 10)
+    {
+      show_congratulations_ui(renderer, game_stat);
+      initiate_player(player, 50, 50);
+      return;
+    }
+
     show_level_transition(renderer, 1, current_level);
     current_level++;
     change_level(current_level);
@@ -87,28 +95,28 @@ void stage1_update(double delta_time)
     {
       initiate_player(player, 570, 70);
     }
-    if (current_level == 5)
+    else if (current_level == 5)
     {
-      initiate_player(player, 50, 300);
+      initiate_player(player, 65, 300);
     }
-    if (current_level == 6)
+    else if (current_level == 6)
     {
       initiate_player(player, 80, 300);
     }
-    if (current_level == 7)
+    else if (current_level == 7)
     {
       initiate_player(player, 650, 50);
     }
-    if (current_level == 8)
+    else if (current_level == 8)
     {
       initiate_player(player, 100, 70);
     }
-    if (current_level == 9)
+    else if (current_level == 9)
     {
       initiate_player(player, 75, 500);
       setup_level_saws(current_level);
     }
-    if (current_level == 10)
+    else if (current_level == 10)
     {
       initiate_player(player, 50, 50);
     }
@@ -133,8 +141,6 @@ void stage1_render(SDL_Renderer *renderer)
 
 
   render_game_ui(renderer, &game_stat);
-
-  SDL_RenderPresent(renderer);
 }
 
 void stage1_cleanup()
